Add -DHXCPP_SCRIPTABLE_REPLACE flag which takes effect when `-D scriptableReplace` is set on haxe
This flag will allow C++ classes to be replaced with cppia versions. This commit does that
by making sure that cppia always overwrites a native class, and by adding a `mReference` class
variable, so that the original static class can be replaced with the script version.

This commit depends on changes from the Haxe repository